requirements.txt
google-api-python-client
python-dotenv
requests
pandas
groq